WIP:Work on deleting external links and breaking links
