+++ src/EPPlus/ExcelAddress.cs
@@ -19,1555 +19,6 @@ using System.Text.RegularExpressions;
 namespace OfficeOpenXml
 {
     /// <summary>
-    /// Extended address information for a table address
-    /// </summary>
-    public class ExcelTableAddress
-    {
-        /// <summary>
-        /// The name of the table
-        /// </summary>
-        public string Name { get; set; }
-        /// <summary>
-        /// Column span
-        /// </summary>
-        public string ColumnSpan { get; set; }
-        /// <summary>
-        /// Reference entire table
-        /// </summary>
-        public bool IsAll { get; set; }
-        /// <summary>
-        /// Reference the table header row
-        /// </summary>
-        public bool IsHeader { get; set; }
-        /// <summary>
-        /// Reference table data
-        /// </summary>
-        public bool IsData { get; set; }
-        /// <summary>
-        /// Reference table totals row
-        /// </summary>
-        public bool IsTotals { get; set; }
-        /// <summary>
-        /// Reference the current table row
-        /// </summary>
-        public bool IsThisRow { get; set; }
-    }
-    /// <summary>
-    /// A range address
-    /// </summary>
-    /// <remarks>Examples of addresses are "A1" "B1:C2" "A:A" "1:1" "A1:E2,G3:G5" </remarks>
-    public class ExcelAddressBase : ExcelCellBase
-    {
-        internal int _fromRow=-1, _toRow, _fromCol, _toCol;
-        internal bool _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed;
-        internal string _wb;
-        internal string _ws;
-        internal string _address;
-
-        internal enum eAddressCollition
-        {
-            No,
-            Partly,
-            Inside,
-            Equal
-        }
-        #region "Constructors"
-        internal ExcelAddressBase()
-        {
-        }
-        /// <summary>
-        /// Creates an Address object
-        /// </summary>
-        /// <param name="fromRow">start row</param>
-        /// <param name="fromCol">start column</param>
-        /// <param name="toRow">End row</param>
-        /// <param name="toColumn">End column</param>
-        public ExcelAddressBase(int fromRow, int fromCol, int toRow, int toColumn)
-        {
-            _fromRow = fromRow;
-            _toRow = toRow;
-            _fromCol = fromCol;
-            _toCol = toColumn;
-            Validate();
-
-            _address = GetAddress(_fromRow, _fromCol, _toRow, _toCol);
-        }
-        /// <summary>
-        /// Creates an Address object
-        /// </summary>
-        /// <param name="worksheetName">Worksheet name</param>
-        /// <param name="fromRow">Start row</param>
-        /// <param name="fromCol">Start column</param>
-        /// <param name="toRow">End row</param>
-        /// <param name="toColumn">End column</param>
-        public ExcelAddressBase(string worksheetName, int fromRow, int fromCol, int toRow, int toColumn)
-        {
-            _ws = worksheetName;
-            _fromRow = fromRow;
-            _toRow = toRow;
-            _fromCol = fromCol;
-            _toCol = toColumn;
-            Validate();
-
-            _address = GetAddress(_fromRow, _fromCol, _toRow, _toCol);
-        }
-
-        internal static bool IsTableAddress(string address)
-        {
-            SplitAddress(address, out string wb, out string ws, out string intAddress);
-            var lPos = intAddress.IndexOf('[');
-            if(lPos >= 0) 
-            {
-                var rPos= intAddress.IndexOf(']',lPos);
-                if(rPos>lPos)
-                {
-                    var c=intAddress[lPos+1];
-                    return !((c >= '0' && c <= '9') || c == '-');
-                }
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// Creates an Address object
-        /// </summary>
-        /// <param name="fromRow">Start row</param>
-        /// <param name="fromCol">Start column</param>
-        /// <param name="toRow">End row</param>
-        /// <param name="toColumn">End column</param>
-        /// <param name="fromRowFixed">Start row fixed</param>
-        /// <param name="fromColFixed">Start column fixed</param>
-        /// <param name="toRowFixed">End row fixed</param>
-        /// <param name="toColFixed">End column fixed</param>
-        public ExcelAddressBase(int fromRow, int fromCol, int toRow, int toColumn, bool fromRowFixed, bool fromColFixed, bool toRowFixed, bool toColFixed) :
-            this(fromRow, fromCol, toRow, toColumn, fromRowFixed, fromColFixed, toRowFixed, toColFixed, null)
-        {
-
-        }
-        internal ExcelAddressBase(int fromRow, int fromCol, int toRow, int toColumn, bool fromRowFixed, bool fromColFixed, bool toRowFixed, bool toColFixed, string worksheetName)
-        {
-            _fromRow = fromRow;
-            _toRow = toRow;
-            _fromCol = fromCol;
-            _toCol = toColumn;
-            _fromRowFixed = fromRowFixed;
-            _fromColFixed = fromColFixed;
-            _toRowFixed = toRowFixed;
-            _toColFixed = toColFixed;
-            _ws = worksheetName;
-            Validate();
-
-            _address = GetAddress(_fromRow, _fromCol, _toRow, _toCol, _fromRowFixed, fromColFixed, _toRowFixed, _toColFixed );
-        }
-        /// <summary>
-        /// Creates an Address object
-        /// </summary>
-        /// <remarks>Examples of addresses are "A1" "B1:C2" "A:A" "1:1" "A1:E2,G3:G5" </remarks>
-        /// <param name="address">The Excel Address</param>
-        /// <ws></ws>
-        public ExcelAddressBase(string address, ExcelWorkbook wb=null, string wsName=null)
-        {
-            SetAddress(address, wb, wsName);
-            if (string.IsNullOrEmpty(_ws)) _ws = wsName;
-        }
-        /// <summary>
-        /// Creates an Address object
-        /// </summary>
-        /// <remarks>Examples of addresses are "A1" "B1:C2" "A:A" "1:1" "A1:E2,G3:G5" </remarks>
-        /// <param name="address">The Excel Address</param>
-        /// <param name="pck">Reference to the package to find information about tables and names</param>
-        /// <param name="referenceAddress">The address</param>
-        public ExcelAddressBase(string address, ExcelPackage pck, ExcelAddressBase referenceAddress)
-        {
-            SetAddress(address, null, null);
-            SetRCFromTable(pck, referenceAddress);
-        }
-
-        internal void SetRCFromTable(ExcelPackage pck, ExcelAddressBase referenceAddress)
-        {
-            if (string.IsNullOrEmpty(_wb) && Table != null)
-            {
-                foreach (var ws in pck.Workbook.Worksheets)
-                {
-                    if (ws is ExcelChartsheet) continue;
-                    foreach (var t in ws.Tables)
-                    {
-                        if (t.Name.Equals(Table.Name, StringComparison.OrdinalIgnoreCase))
-                        {
-                            _ws = ws.Name;
-                            if (Table.IsAll)
-                            {
-                                _fromRow = t.Address._fromRow;
-                                _toRow = t.Address._toRow;
-                            }
-                            else
-                            {
-                                if (Table.IsThisRow)
-                                {
-                                    if (referenceAddress == null)
-                                    {
-                                        _fromRow = -1;
-                                        _toRow = -1;
-                                    }
-                                    else
-                                    {
-                                        _fromRow = referenceAddress._fromRow;
-                                        _toRow = _fromRow;
-                                    }
-                                }
-                                else if (Table.IsHeader && Table.IsData)
-                                {
-                                    _fromRow = t.Address._fromRow;
-                                    _toRow = t.ShowTotal ? t.Address._toRow - 1 : t.Address._toRow;
-                                }
-                                else if (Table.IsData && Table.IsTotals)
-                                {
-                                    _fromRow = t.ShowHeader ? t.Address._fromRow + 1 : t.Address._fromRow;
-                                    _toRow = t.Address._toRow;
-                                }
-                                else if (Table.IsHeader)
-                                {
-                                    _fromRow = t.ShowHeader ? t.Address._fromRow : -1;
-                                    _toRow = t.ShowHeader ? t.Address._fromRow : -1;
-                                }
-                                else if (Table.IsTotals)
-                                {
-                                    _fromRow = t.ShowTotal ? t.Address._toRow : -1;
-                                    _toRow = t.ShowTotal ? t.Address._toRow : -1;
-                                }
-                                else
-                                {
-                                    _fromRow = t.ShowHeader ? t.Address._fromRow + 1 : t.Address._fromRow;
-                                    _toRow = t.ShowTotal ? t.Address._toRow - 1 : t.Address._toRow;
-                                }
-                            }
-
-                            if (string.IsNullOrEmpty(Table.ColumnSpan))
-                            {
-                                _fromCol = t.Address._fromCol;
-                                _toCol = t.Address._toCol;
-                                return;
-                            }
-                            else
-                            {
-                                var col = t.Address._fromCol;
-                                var cols = Table.ColumnSpan.Split(':');
-                                foreach (var c in t.Columns)
-                                {
-                                    if (_fromCol <= 0 && cols[0].Equals(c.Name, StringComparison.OrdinalIgnoreCase))   //Issue15063 Add invariant igore case
-                                    {
-                                        _fromCol = col;
-                                        if (cols.Length == 1)
-                                        {
-                                            _toCol = _fromCol;
-                                            return;
-                                        }
-                                    }
-                                    else if (cols.Length > 1 && _fromCol > 0 && cols[1].Equals(c.Name, StringComparison.OrdinalIgnoreCase)) //Issue15063 Add invariant igore case
-                                    {
-                                        _toCol = col;
-                                        return;
-                                    }
-
-                                    col++;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        internal ExcelAddressBase Intersect(ExcelAddressBase address)
-        {
-            if(address._fromRow > _toRow || _toRow < address._fromRow ||
-               address._fromCol > _toCol || _toCol < address._fromCol)
-            {
-                return null;
-            }
-            
-            var fromRow = Math.Max(address._fromRow, _fromRow);
-            var toRow = Math.Min(address._toRow, _toRow);
-            var fromCol = Math.Max(address._fromCol, _fromCol);
-            var toCol = Math.Min(address._toCol, _toCol);
-
-            return new ExcelAddressBase(fromRow, fromCol, toRow, toCol);
-        }
-
-        internal bool IsInside(ExcelAddressBase effectedAddress)
-        {
-            var c = Collide(effectedAddress);
-            return c == ExcelAddressBase.eAddressCollition.Equal ||
-                   c == ExcelAddressBase.eAddressCollition.Inside;
-        }
-        /// <summary>
-        /// Address is an defined name
-        /// </summary>
-        /// <param name="address">the name</param>
-        /// <param name="isName">Should always be true</param>
-        internal ExcelAddressBase(string address, bool isName)
-        {
-            if (isName)
-            {
-                _address = address;
-                _fromRow = -1;
-                _fromCol = -1;
-                _toRow = -1;
-                _toCol = -1;
-                _start = null;
-                _end = null;
-            }
-            else
-            {
-                SetAddress(address, null, null);
-            }
-        }
-        /// <summary>
-        /// Sets the address
-        /// </summary>
-        /// <param name="address">The address</param>
-        /// <param name="wb"></param>
-        /// <param name="wsName"></param>
-        protected internal void SetAddress(string address, ExcelWorkbook wb, string wsName)
-        {
-            address = address.Trim();
-            if (Utils.ConvertUtil._invariantCompareInfo.IsPrefix(address, "'") || Utils.ConvertUtil._invariantCompareInfo.IsPrefix(address, "["))
-            {
-                SetWbWs(address);
-            }
-            else
-            {
-                _address = address;
-            }
-
-            _addresses = null;
-            if (_address.IndexOfAny(new char[] {',','!', '['}) > -1)
-            {
-                //Advanced address. Including Sheet or multi or table.
-                ExtractAddress(_address);
-            }
-            else
-            {
-                //Simple address
-                GetRowColFromAddress(_address, out _fromRow, out _fromCol, out _toRow, out  _toCol, out _fromRowFixed, out _fromColFixed,  out _toRowFixed, out _toColFixed, wb, wsName);
-                _start = null;
-                _end = null;
-            }
-            _address = address;
-            Validate();
-        }
-        /// <summary>
-        /// Called when the address changes
-        /// </summary>
-        internal protected virtual void ChangeAddress()
-        {
-        }
-        private void SetWbWs(string address)
-        {
-            int pos;
-            if (address[0] == '[')
-            {
-                pos = address.IndexOf(']');
-                _wb = address.Substring(1, pos - 1);                
-                _ws = address.Substring(pos + 1);                
-            }
-            else
-            {
-                _wb = "";
-                _ws = address;
-            }
-            if(_ws.StartsWith("'", StringComparison.OrdinalIgnoreCase))
-            {
-                pos = _ws.IndexOf("'",1, StringComparison.OrdinalIgnoreCase);
-                while(pos>0 && pos+1<_ws.Length && _ws[pos+1]=='\'')
-                {
-                    _ws = _ws.Substring(0, pos) + _ws.Substring(pos+1);
-                    pos = _ws.IndexOf("'", pos+1, StringComparison.OrdinalIgnoreCase);
-                }
-                if (pos>0)
-                {
-                    if(_ws.Length-1==pos)
-                    {
-                        _address = "A:XFD";
-                    }
-                    else if (_ws[pos+1]!='!')
-                    {
-                        throw new InvalidOperationException($"Address is not valid {address}. Missing ! after sheet name.");
-                    }
-                    else
-                    {
-                        _address = _ws.Substring(pos + 2);
-                    }
-                    _ws = _ws.Substring(1, pos-1);
-                    pos = _address.IndexOf(":'", StringComparison.OrdinalIgnoreCase);
-                    if(pos>0)
-                    {
-                        var a1 = _address.Substring(0,pos);
-                        pos = _address.LastIndexOf("\'!", StringComparison.OrdinalIgnoreCase);
-                        if (pos > 0)
-                        {
-                            var a2 = _address.Substring(pos+2);
-                            _address=a1 + ":" + a2; //Remove any worksheet on second reference of the address. 
-                        }
-                    }
-                    return;
-                }
-            }
-            pos = _ws.IndexOf('!');
-
-            if (pos==0)
-            {
-                _address = _ws.Substring(1);
-                _ws = _wb;
-                _wb = "";
-            }
-            else if (pos > -1)
-            {
-                _address = _ws.Substring(pos + 1);
-                _ws = _ws.Substring(0, pos);
-            }
-            else
-            {
-                _address = address;
-            }
-            if(string.IsNullOrEmpty(_address))
-            {
-                _address = "A:XFD";
-            }
-        }
-        internal void ChangeWorksheet(string wsName, string newWs)
-        {
-            if (_ws == wsName) _ws = newWs;
-            var fullAddress = GetAddress();
-            
-            if (Addresses != null)
-            {
-                foreach (var a in Addresses)
-                {
-                    if (a._ws == wsName)
-                    {
-                        a._ws = newWs;
-                        fullAddress += "," + a.GetAddress();
-                    }
-                    else
-                    {
-                        fullAddress += "," + a._address;
-                    }
-                }
-            }
-            _address = fullAddress;
-        }
-
-        private string GetAddress()
-        {
-            string address = GetAddressWorkBookWorkSheet();
-            if (IsName)
-                return address + GetAddress(_fromRow, _fromCol, _toRow, _toCol);
-            else
-                return address + GetAddress(_fromRow, _fromCol, _toRow, _toCol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed);
-        }
-
-        internal string GetAddressWorkBookWorkSheet()
-        {
-            var address = "";
-
-            if (string.IsNullOrEmpty(_ws) == false)
-            {
-                if (string.IsNullOrEmpty(_wb) == false)
-                {
-                    address = "[" + _wb + "]";
-                }
-
-                if (_address.IndexOf("'!", StringComparison.OrdinalIgnoreCase) >=0 || ExcelWorksheet.NameNeedsApostrophes(_ws))
-                {
-                    address += string.Format("'{0}'!", _ws.Replace("'","''"));
-                }
-                else
-                {
-                    address += string.Format("{0}!", _ws);
-                }
-            }
-
-            return address;
-        }
-        #endregion
-        internal ExcelCellAddress _start = null;
-        /// <summary>
-        /// Gets the row and column of the top left cell.
-        /// </summary>
-        /// <value>The start row column.</value>
-        public ExcelCellAddress Start
-        {
-            get
-            {
-                if (_start == null)
-                {
-                    _start = new ExcelCellAddress(_fromRow, _fromCol, _fromRowFixed, _fromColFixed);
-                }
-                return _start;
-            }
-        }
-        internal ExcelCellAddress _end = null;
-        /// <summary>
-        /// Gets the row and column of the bottom right cell.
-        /// </summary>
-        /// <value>The end row column.</value>
-        public ExcelCellAddress End
-        {
-            get
-            {
-                if (_end == null)
-                {
-                    _end = new ExcelCellAddress(_toRow, _toCol, _toRowFixed, _toColFixed);
-                }
-                return _end;
-            }
-        }
-        internal ExcelTableAddress _table = null;
-        /// <summary>
-        /// If the address is refering a table, this property contains additional information 
-        /// </summary>
-        public ExcelTableAddress Table
-        {
-            get
-            {
-                return _table;
-            }
-        }
-
-        /// <summary>
-        /// The address for the range
-        /// </summary>
-        public virtual string Address
-        {
-            get
-            {
-                return _address;
-            }
-        }
-        /// <summary>
-        /// The full address including the worksheet
-        /// </summary>
-        public string FullAddress
-        {
-            get
-            {
-                string a="";
-                if(_addresses != null)
-                {
-                    foreach(var sa in _addresses)
-                    {
-                        a += ","+sa.GetAddress();
-                    }
-                }
-                else
-                {
-                    a = GetAddress();
-                }
-                return a;
-            }
-        }
-        /// <summary>
-        /// If the address is a defined name
-        /// </summary>
-        public bool IsName
-        {
-            get
-            {
-                return _fromRow < 0;
-            }
-        }
-        /// <summary>
-        /// Returns the address text
-        /// </summary>
-        /// <returns></returns>
-        public override string ToString()
-        {
-            return _address;
-        }
-        string _firstAddress;
-        /// <summary>
-        /// returns the first address if the address is a multi address.
-        /// A1:A2,B1:B2 returns A1:A2
-        /// </summary>
-        internal string FirstAddress
-        {
-            get
-            {
-                if (string.IsNullOrEmpty(_firstAddress))
-                {
-                    return _address;
-                }
-                else
-                {
-                    return _firstAddress;
-                }
-            }
-        }
-        internal string AddressSpaceSeparated
-        {
-            get
-            {
-                return _address.Replace(',', ' '); //Conditional formatting and a few other places use space as separator for mulit addresses.
-            }
-        }
-        /// <summary>
-        /// Validate the address
-        /// </summary>
-        protected void Validate()
-        {
-            if (_fromRow > _toRow || _fromCol > _toCol)
-            {
-                throw new ArgumentOutOfRangeException("Start cell Address must be less or equal to End cell address");
-            }
-        }
-        internal string WorkSheetName
-        {
-            get
-            {
-                return _ws;
-            }
-        }
-        internal List<ExcelAddress> _addresses = null;
-        internal virtual List<ExcelAddress> Addresses
-        {
-            get
-            {
-                return _addresses;
-            }
-        }
-
-        private bool ExtractAddress(string fullAddress)
-        {
-            var brackPos=new Stack<int>();
-            var bracketParts=new List<string>();
-            string first="", second="";
-            bool isText=false, hasSheet=false;
-            _addresses = null;
-            try
-            {
-                if (fullAddress == "#REF!")
-                {
-                    SetAddress(ref fullAddress, ref second, ref hasSheet);
-                    return true;
-                }
-                else if (Utils.ConvertUtil._invariantCompareInfo.IsPrefix(fullAddress, "!"))
-                {
-                    // invalid address!
-                    return false;
-                }
-                for (int i = 0; i < fullAddress.Length; i++)
-                {
-                    var c = fullAddress[i];
-                    if (c == '\'')
-                    {
-                        if (isText && i + 1 < fullAddress.Length && fullAddress[i] == '\'')
-                        {
-                            if (hasSheet)
-                            {
-                                second += c;
-                            }
-                            else
-                            {
-                                first += c;
-                            }
-                        }
-                        isText = !isText;
-                    }
-                    else
-                    {
-                        if (brackPos.Count > 0)
-                        {
-                            if (c == '[' && !isText)
-                            {
-                                brackPos.Push(i);
-                            }
-                            else if (c == ']' && !isText)
-                            {
-                                if (brackPos.Count > 0)
-                                {
-                                    var from = brackPos.Pop();
-                                    bracketParts.Add(fullAddress.Substring(from + 1, i - from - 1));
-
-                                    if (brackPos.Count == 0)
-                                    {
-                                        HandleBrackets(first, second, bracketParts);
-                                    }
-                                }
-                                else
-                                {
-                                    //Invalid address!
-                                    return false;
-                                }
-                            }
-                        }
-                        else if (c == '[' && !isText)
-                        {
-                            brackPos.Push(i);
-                        }
-                        else if (c == '!' && !isText && !first.EndsWith("#REF") && !second.EndsWith("#REF"))
-                        {
-                            // the following is to handle addresses that specifies the
-                            // same worksheet twice: Sheet1!A1:Sheet1:A3
-                            // They will be converted to: Sheet1!A1:A3
-                            if (hasSheet && second != null && second.ToLower().EndsWith(first.ToLower()))
-                            {
-                                second = Regex.Replace(second, $"{first}$", string.Empty);
-                            }
-                            hasSheet = true;
-                        }
-                        else if (c == ',' && !isText)
-                        {
-                            if(_addresses==null) _addresses = new List<ExcelAddress>();
-                            SetAddress(ref first, ref second, ref hasSheet);
-                        }
-                        else
-                        {
-                            if (hasSheet)
-                            {
-                                second += c;
-                            }
-                            else
-                            {
-                                first += c;
-                            }
-                        }
-                    }
-                }
-                if (Table == null)
-                {
-                    SetAddress(ref first, ref second, ref hasSheet);
-                }
-                return true;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        private void HandleBrackets(string first, string second, List<string> bracketParts)
-        {
-            if(!string.IsNullOrEmpty(first))
-            {
-                _table = new ExcelTableAddress();
-                Table.Name = first;
-                foreach (var s in bracketParts)
-                {
-                    if(s.IndexOf('[')<0)
-                    {
-                        switch(s.ToLower(CultureInfo.InvariantCulture))                
-                        {
-                            case "#all":
-                                _table.IsAll = true;
-                                break;
-                            case "#headers":
-                               _table.IsHeader = true;
-                                break;
-                            case "#data":
-                                _table.IsData = true;
-                                break;
-                            case "#totals":
-                                _table.IsTotals = true;
-                                break;
-                            case "#this row":
-                                _table.IsThisRow = true;
-                                break;
-                            default:
-                                if(string.IsNullOrEmpty(_table.ColumnSpan))
-                                {
-                                    _table.ColumnSpan=s;
-                                }
-                                else
-                                {
-                                    _table.ColumnSpan += ":" + s;
-                                }
-                                break;
-                        }                
-                    }
-                }
-            }
-        }
-        #region Address manipulation methods
-        internal eAddressCollition Collide(ExcelAddressBase address, bool ignoreWs=false)
-        {
-            if (ignoreWs == false && address.WorkSheetName != WorkSheetName && address.WorkSheetName != null && WorkSheetName!=null)
-            {
-                return eAddressCollition.No;
-            }
-
-            return Collide(address._fromRow, address._fromCol, address._toRow, address._toCol);
-        }
-        internal eAddressCollition Collide(int row, int col)
-        {
-            return Collide(row, col, row, col);
-        }
-        internal eAddressCollition Collide(int fromRow, int fromCol, int toRow, int toCol)
-        {
-            if (DoNotCollide(fromRow, fromCol, toRow, toCol))
-            {
-                return eAddressCollition.No;
-            }
-            else if (fromRow == _fromRow && fromCol == _fromCol &&
-                    toRow == _toRow && toCol == _toCol)
-            {
-                return eAddressCollition.Equal;
-            }
-            else if (fromRow >= _fromRow && toRow <= _toRow &&
-                     fromCol >= _fromCol && toCol <= _toCol)
-            {
-                return eAddressCollition.Inside;
-            }
-            else
-                return eAddressCollition.Partly;
-        }
-
-        internal bool DoNotCollide(int fromRow, int fromCol, int toRow, int toCol)
-        {
-            return fromRow > _toRow || fromCol > _toCol
-                   ||
-                   _fromRow > toRow || _fromCol > toCol;
-        }
-
-        internal bool CollideFullRowOrColumn(ExcelAddressBase address)
-        {
-            return CollideFullRowOrColumn(address._fromRow, address._fromCol, address._toRow, address._toCol);
-        }
-        internal bool CollideFullRowOrColumn(int fromRow, int fromCol, int toRow, int toCol)
-        {
-            return (CollideFullRow(fromRow, toRow) && CollideColumn(fromCol, toCol)) || 
-                   (CollideFullColumn(fromCol, toCol) && CollideRow(fromRow, toRow));
-        }
-        private bool CollideColumn(int fromCol, int toCol)
-        {
-            return fromCol  <= _toCol && toCol >= _fromCol;
-        }
-
-        internal bool CollideRow(int fromRow, int toRow)
-        {
-            return fromRow <= _toRow && toRow >= _fromRow;
-        }
-        internal bool CollideFullRow(int fromRow, int toRow)
-        {
-            return fromRow <= _fromRow && toRow >= _toRow;
-        }
-        internal bool CollideFullColumn(int fromCol, int toCol)
-        {
-            return fromCol <= _fromCol && toCol >= _toCol;
-        }
-        internal ExcelAddressBase AddRow(int row, int rows, bool setFixed=false)
-        {
-            if (row > _toRow)
-            {
-                return this;
-            }
-            else if (row <= _fromRow)
-            {
-                return new ExcelAddressBase((setFixed && _fromRowFixed ? _fromRow : _fromRow + rows), _fromCol, (setFixed && _toRowFixed ? _toRow : _toRow + rows), _toCol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, WorkSheetName);
-            }
-            else
-            {
-                return new ExcelAddressBase(_fromRow, _fromCol, (setFixed && _toRowFixed ? _toRow : _toRow + rows), _toCol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, WorkSheetName);
-            }
-        }
-        internal ExcelAddressBase DeleteRow(int row, int rows, bool setFixed = false)
-        {
-            if (row > _toRow) //After
-            {
-                return this;
-            }            
-            else if (row+rows < _fromRow || (_fromRowFixed && row < _fromRow)) //Before
-            {
-                return new ExcelAddressBase((setFixed && _fromRowFixed ? _fromRow : _fromRow - rows), _fromCol, (setFixed && _toRowFixed ? _toRow : _toRow - rows), _toCol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, WorkSheetName);
-            }
-            else if (row <= _fromRow && row + rows > _toRow) //Inside
-            {
-                return null;
-            }
-            else  //Partly
-            {
-                if (row <= _fromRow)
-                {
-                    return new ExcelAddressBase(row, _fromCol, (setFixed && _toRowFixed ? _toRow : _toRow - rows), _toCol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, WorkSheetName);
-                }
-                else
-                {
-                    return new ExcelAddressBase(_fromRow, _fromCol, (setFixed && _toRowFixed ? _toRow : _toRow - rows < row ? row - 1 : _toRow - rows), _toCol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, WorkSheetName);
-                }
-            }
-        }
-        internal ExcelAddressBase AddColumn(int col, int cols, bool setFixed = false)
-        {
-            if (col > _toCol)
-            {
-                return this;
-            }
-            else if (col <= _fromCol)
-            {
-                return new ExcelAddressBase(_fromRow, (setFixed && _fromColFixed ? _fromCol : _fromCol + cols), _toRow, (setFixed && _toColFixed ? _toCol : _toCol + cols), _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, WorkSheetName);
-            }
-            else
-            {
-                return new ExcelAddressBase(_fromRow, _fromCol, _toRow, (setFixed && _toColFixed ? _toCol : _toCol + cols), _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, WorkSheetName);
-            }
-        }
-        internal ExcelAddressBase DeleteColumn(int col, int cols, bool setFixed = false)
-        {
-            if (col > _toCol) //After
-            {
-                return this;
-            }
-            else if (col + cols < _fromCol || (_fromColFixed && col < _fromCol)) //Before
-            {
-                return new ExcelAddressBase(_fromRow, (setFixed && _fromColFixed ? _fromCol : _fromCol - cols), _toRow, (setFixed && _toColFixed ? _toCol :_toCol - cols), _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, WorkSheetName);
-            }
-            else if (col <= _fromCol && col + cols > _toCol) //Inside
-            {
-                return null;
-            }
-            else  //Partly
-            {
-                if (col <= _fromCol)
-                {
-                    return new ExcelAddressBase(_fromRow, col, _toRow, (setFixed && _toColFixed ? _toCol : _toCol - cols), _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, _ws);
-                }
-                else
-                {
-                    return new ExcelAddressBase(_fromRow, _fromCol, _toRow, (setFixed && _toColFixed ? _toCol :_toCol - cols < col ? col - 1 : _toCol - cols), _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed, _ws);
-                }
-            }
-        }
-        internal ExcelAddressBase Insert(ExcelAddressBase address, eShiftTypeInsert Shift)
-        {
-            //Before or after, no change
-            if(_toRow < address._fromRow || _toCol < address._fromCol || (_fromRow > address._toRow && _fromCol > address._toCol))
-            {
-                return this;
-            }
-
-            int rows = address.Rows;
-            int cols = address.Columns;
-            string retAddress = "";
-            if (Shift==eShiftTypeInsert.Right)
-            {
-                if (address._fromRow > _fromRow)
-                {
-                    retAddress = GetAddress(_fromRow, _fromCol, address._fromRow, _toCol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed);
-                }
-                if(address._fromCol > _fromCol)
-                {
-                    retAddress = GetAddress(_fromRow < address._fromRow ? _fromRow : address._fromRow, _fromCol, address._fromRow, _toCol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed);
-                }
-            }
-            if (_toRow < address._fromRow)
-            {
-                if (_fromRow < address._fromRow)
-                {
-
-                }
-                else
-                {
-                }
-            }
-            return null;
-        }
-        #endregion
-        private void SetAddress(ref string first, ref string second, ref bool hasSheet)
-        {
-            string ws, address;
-            if (hasSheet)
-            {
-                ws = first;
-                address = second;
-                first = "";
-                second = "";
-            }
-            else
-            {
-                address = first;
-                ws = "";
-                first = "";
-            }
-            hasSheet = false;
-            if (string.IsNullOrEmpty(_firstAddress))
-            {
-                if (string.IsNullOrEmpty(_ws) || !string.IsNullOrEmpty(ws))
-                {
-                    _ws = ws;
-                    
-                }
-                _firstAddress = address;
-                GetRowColFromAddress(address, out _fromRow, out _fromCol, out _toRow, out  _toCol, out _fromRowFixed, out _fromColFixed, out _toRowFixed, out _toColFixed);
-                _start = null;
-                _end = null;
-            }
-            if (_addresses != null)
-            {
-                _addresses.Add(new ExcelAddress(_ws, address));
-            }
-        }
-        internal enum AddressType
-        {
-            Invalid,
-            InternalAddress,
-            ExternalAddress,
-            InternalName,
-            ExternalName,
-            Formula,
-            R1C1
-        }
-
-        internal static AddressType IsValid(string Address, bool r1c1=false)
-        {
-            double d;
-            if (Address == "#REF!")
-            {
-                return AddressType.Invalid;
-            }
-            else if(double.TryParse(Address, NumberStyles.Any, CultureInfo.InvariantCulture, out d)) //A double, no valid address
-            {
-                return AddressType.Invalid;
-            }
-            else if (IsFormula(Address))
-            {
-                return AddressType.Formula;
-            }
-            
-            else
-            {
-                if (r1c1 && IsR1C1(Address))
-                {
-                    return AddressType.R1C1;
-                }
-                else
-                {
-                    string wb, ws, intAddress;
-                    if (SplitAddress(Address, out wb, out ws, out intAddress))
-                    {
-                        if (Address.Contains("[")) //Table reference
-                        {
-                            return string.IsNullOrEmpty(wb) ? AddressType.InternalAddress : AddressType.ExternalAddress;
-                        }
-                        else if (intAddress.Contains(","))
-                        {
-                            intAddress = intAddress.Substring(0, intAddress.IndexOf(','));
-                        }
-                        if (IsAddress(intAddress))
-                        {
-                            return string.IsNullOrEmpty(wb) ? AddressType.InternalAddress : AddressType.ExternalAddress;
-                        }
-                        else
-                        {
-                            return string.IsNullOrEmpty(wb) ? AddressType.InternalName : AddressType.ExternalName;
-                        }
-                    }
-                    else
-                    {
-                        return AddressType.Invalid;
-                    }
-                }
-            }
-        }
-
-        private static bool IsR1C1(string address)
-        {
-            var start = address.LastIndexOf("!", address.Length-1, StringComparison.OrdinalIgnoreCase);
-            if (start>=0)
-            {
-                address = address.Substring(start + 1);
-            }
-            address = address.ToUpper();
-            if (string.IsNullOrEmpty(address) || (address[0]!='R' && address[0]!='C'))
-            {
-                return false;
-            }
-            bool isC = false, isROrC = false;
-            bool startBracket = false;
-            foreach(var c in address)
-            {
-                switch(c)
-                {
-                    case 'C':
-                        isC = true;
-                        isROrC = true;
-                        break;
-                    case 'R':
-                        if (isC)
-                            return false;
-                        isROrC = true;
-                        break;
-                    case '[':
-                        startBracket = true;
-                        break;
-                    case ']':
-                        if (startBracket == false) return false;
-                        isROrC = false;
-                        break;
-                    case ':':
-                        isC = false;
-                        startBracket = false;
-                        isROrC = false;
-                        break;
-                    default:
-                        if((c>='0' && c<='9') ||c=='-')
-                        {
-                            if(isROrC==false)
-                            {
-                                return false;
-                            }
-                        }
-                        else
-                        {
-                            return false;
-                        }
-                        break;
-                }
-            }
-            return true;
-        }
-
-        private static bool IsAddress(string intAddress)
-        {
-            if(string.IsNullOrEmpty(intAddress)) return false;            
-            var cells = intAddress.Split(':');
-            int fromRow, toRow, fromCol, toCol;
-
-            if(!GetRowCol(cells[0], out fromRow, out fromCol, false))
-            {
-                return false;
-            }
-            if (cells.Length > 1)
-            {
-                if (!GetRowCol(cells[1], out toRow, out toCol, false))
-                {
-                    return false;
-                }
-            }
-            else
-            {
-                toRow = fromRow;
-                toCol = fromCol;
-            }
-            if( fromRow <= toRow && 
-                fromCol <= toCol && 
-                fromCol > -1 && 
-                toCol <= ExcelPackage.MaxColumns && 
-                fromRow > -1 && 
-                toRow <= ExcelPackage.MaxRows)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        private static bool SplitAddress(string Address, out string wb, out string ws, out string intAddress)
-        {
-            wb = "";
-            ws = "";
-            intAddress = "";
-            var text = "";
-            bool isText = false;
-            var brackPos=-1;
-            for (int i = 0; i < Address.Length; i++)
-            {
-                if (Address[i] == '\'')
-                {
-                    isText = !isText;
-                    if(i>0 && Address[i-1]=='\'')
-                    {
-                        text += "'";
-                    }
-                }
-                else
-                {
-                    if(Address[i]=='!' && !isText)
-                    {
-                        if (text.Length>0 && text[0] == '[')
-                        {
-                            wb = text.Substring(1, text.IndexOf(']') - 1);
-                            ws = text.Substring(text.IndexOf(']') + 1);
-                        }
-                        else
-                        {
-                            ws=text;
-                        }
-                        intAddress=Address.Substring(i+1);
-                        return true;
-                    }
-                    else
-                    {
-                        if(Address[i]=='[' && !isText)
-                        {
-                            if (i > 0) //Table reference return full address;
-                            {
-                                intAddress=Address;
-                                return true;
-                            }
-                            brackPos=i;
-                        }
-                        else if(Address[i]==']' && !isText)
-                        {
-                            if (brackPos > -1)
-                            {
-                                wb = text;
-                                text = "";
-                            }
-                            else
-                            {
-                                return false;
-                            }
-                        }
-                        else
-                        {
-                            text+=Address[i];
-                        }
-                    }
-                }
-            }
-            intAddress = text;
-            return true;
-        }
-
-        private static readonly HashSet<char> _tokens = new HashSet<char>(new char[] { '+', '-', '*', '/', '^', '&', '=', '<', '>', '(', ')', '{', '}', '%', '\"' }); //See TokenSeparatorProvider
-        internal static bool IsFormula(string address)
-        {
-            var isText = false;
-            var tableNameCount = 0;
-            for (int i = 0; i < address.Length; i++)
-            {
-                var addressChar = address[i];
-                if (addressChar == '\'')
-                {
-                    isText = !isText;
-                }
-                else if (isText == false && addressChar == '[')
-                    tableNameCount++;
-                else if (isText == false && addressChar == ']')
-                    tableNameCount--;
-                else if(tableNameCount==0)
-                {
-                    if (isText == false && _tokens.Contains(addressChar))
-                    {
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-
-        private static bool IsValidName(string address)
-        {
-            if (Regex.IsMatch(address, "[^0-9./*-+,½!\"@#£%&/{}()\\[\\]=?`^~':;<>|][^/*-+,½!\"@#£%&/{}()\\[\\]=?`^~':;<>|]*"))
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Number of rows int the address
-        /// </summary>
-        public int Rows 
-        {
-            get
-            {
-                return _toRow - _fromRow+1;
-            }
-        }
-        /// <summary>
-        /// Number of columns int the address
-        /// </summary>
-        public int Columns
-        {
-            get
-            {
-                return _toCol - _fromCol + 1;
-            }
-        }
-        /// <summary>
-        /// Returns true if the range spans a full row
-        /// </summary>
-        /// <returns></returns>
-        public bool IsFullRow
-        {
-            get
-            {
-                return _fromCol == 1 && _toCol == ExcelPackage.MaxColumns;
-            }
-        }
-        /// <summary>
-        /// Returns true if the range spans a full column
-        /// </summary>
-        /// <returns></returns>
-        public bool IsFullColumn
-        {
-            get
-            {
-                return _fromRow == 1 && _toRow == ExcelPackage.MaxRows;
-            }
-        }
-
-        internal bool IsSingleCell
-        {
-            get
-            {
-                return (_fromRow == _toRow && _fromCol == _toCol);
-            }
-        }
-
-        public string LocalAddress 
-        { 
-            get
-            {
-                var ix = _address.TrimEnd().LastIndexOf('!', _address.Length - 2);  //Last index can be ! if address is #REF!, so check from 
-                if (ix>=0)
-                {
-                    return _address.Substring(ix + 1);
-                }
-                return _address;
-            }
-        }
-
-        internal static String GetWorkbookPart(string address)
-        {
-            var ix = 0;
-            if (address[0] == '[')
-            {
-                ix = address.IndexOf(']') + 1;
-                if (ix > 0)
-                {
-                    return address.Substring(1, ix - 2);
-                }
-            }
-            return "";
-        }
-        internal static string GetWorksheetPart(string address, string defaultWorkSheet)
-        {
-            int ix=0;
-            return GetWorksheetPart(address, defaultWorkSheet, ref ix);
-        }
-        internal static string GetWorksheetPart(string address, string defaultWorkSheet, ref int endIx)
-        {
-            if(address=="") return defaultWorkSheet;
-            var ix = 0;
-            if (address[0] == '[')
-            {
-                ix = address.IndexOf(']')+1;
-            }
-            if (ix > 0 && ix < address.Length)
-            {
-                if (address[ix] == '\'')
-                {
-                    return GetString(address, ix, out endIx);
-                }
-                else
-                {
-                    var ixEnd = address.IndexOf('!',ix);
-                    if(ixEnd>ix)
-                    {
-                        return address.Substring(ix, ixEnd-ix);
-                    }
-                    else
-                    {
-                        return defaultWorkSheet;
-                    }
-                }
-            }
-            else
-            {
-                return defaultWorkSheet;
-            }
-        }
-        internal static string GetAddressPart(string address)
-        {
-            var ix=0;
-            GetWorksheetPart(address, "", ref ix);
-            if(ix<address.Length)
-            {
-                if (address[ix] == '!')
-                {
-                    return address.Substring(ix + 1);
-                }
-                else
-                {
-                    return "";
-                }
-            }
-            else
-            {
-                return "";
-            }
-
-        }
-        internal static void SplitAddress(string fullAddress, out string wb, out string ws, out string address, string defaultWorksheet="")
-        {
-            wb = GetWorkbookPart(fullAddress);
-            int ix=0;
-            ws = GetWorksheetPart(fullAddress, defaultWorksheet, ref ix);
-            if (ix < fullAddress.Length)
-            {
-                if (fullAddress[ix] == '!')
-                {
-                    address = fullAddress.Substring(ix + 1);
-                }
-                else
-                {
-                    address = fullAddress.Substring(ix);
-                }
-            }
-            else
-            {
-                address="";
-            }
-        }
-        internal static List<string[]> SplitFullAddress(string fullAddress)
-        {
-            var addresses = new List<string[]>();
-            var currentAddress = new string[3];
-            bool isInWorkbook = false;
-            bool isInWorksheet = false;
-            bool isInAddress = false;
-            bool isInText = false;
-            var prevPos = 0;
-            for (int i=0;i<fullAddress.Length;i++)
-            {
-                if (isInWorkbook == false &&
-                    isInWorksheet == false &&
-                    isInAddress == false)
-                {
-                    if (fullAddress[i] == '[')
-                    {
-                        isInWorkbook = true;
-                        prevPos = i + 1;
-                    }
-                    else if(fullAddress[i] == '\'')
-                    {
-                        isInWorksheet = true;
-                        isInText = true;
-                        prevPos = i + 1;
-                    }
-                    else if (fullAddress[i]=='!')
-                    {
-                        isInAddress = true;
-                        prevPos = i + 1;
-                    }
-                    else
-                    {
-                        isInAddress = true;
-                    }
-                }
-                else if(isInWorkbook)
-                {
-                    if (fullAddress[i] == ']')
-                    {
-                        currentAddress[0] = fullAddress.Substring(prevPos, i - prevPos);
-                        isInWorkbook = false;
-                    }
-                }
-                else if(isInWorksheet)
-                {
-                    if (fullAddress[i] == '\'')
-                    {
-                        isInText = !isInText;
-                    }
-                    else if (isInText==false && fullAddress[i] == '!')
-                    {
-                        currentAddress[1] = fullAddress.Substring(prevPos, i -prevPos - 1).Replace("''","'");
-                        prevPos = i + 1;
-                        isInWorksheet = false;
-                    }
-                }
-                else if(isInAddress)
-                {
-                    if(fullAddress[i] == '!')
-                    {
-                        currentAddress[1] = fullAddress.Substring(prevPos, i - prevPos);
-                        prevPos = i + 1;
-                    }
-                    else if (fullAddress[i]==',')
-                    {
-                        currentAddress[2] = fullAddress.Substring(prevPos, i - prevPos);
-                        addresses.Add(currentAddress);
-                        prevPos = i + 1;
-                        isInAddress = false;
-                        currentAddress = new string[3];
-                    }
-                }
-            }
-
-            if(isInWorkbook || isInWorksheet)
-            {
-                throw (new ArgumentException($"Invalid address {fullAddress}"));
-            }
-            currentAddress[2] = fullAddress.Substring(prevPos, fullAddress.Length - prevPos);
-            addresses.Add(currentAddress);
-            return addresses;
-        }
-        private static string GetString(string address, int ix, out int endIx)
-        {
-            var strIx = address.IndexOf("''", StringComparison.OrdinalIgnoreCase);
-            var prevStrIx = ix;
-            while(strIx > -1) 
-            {
-                prevStrIx = strIx;
-                strIx = address.IndexOf("''", StringComparison.OrdinalIgnoreCase);
-            }
-            endIx = address.IndexOf("'");
-            return address.Substring(ix, endIx - ix).Replace("''","'");
-        }
-
-        internal bool IsValidRowCol()
-        {
-            return !(_fromRow > _toRow  ||
-                   _fromCol > _toCol ||
-                   _fromRow < 1 ||
-                   _fromCol < 1 ||
-                   _toRow > ExcelPackage.MaxRows ||
-                   _toCol > ExcelPackage.MaxColumns);
-        }
-        public override bool Equals(object obj)
-        {
-            if (obj is ExcelAddressBase a)
-            {
-                if (Addresses==null || a.Addresses==null)
-                {
-                    if (Addresses?.Count > 1 || a.Addresses?.Count > 1) return false;
-                    return IsEqual(this, a);
-                }
-                else
-                {
-                    for(int i=0;i<Addresses.Count;i++)
-                    {
-                        if (IsEqual(Addresses[i], a.Addresses[i]) == false)
-                        {
-                            return false;
-                        }
-                    }
-                    return true;
-                }
-            }
-            else
-            {
-                return _address == obj.ToString();
-            }
-        }
-
-        private bool IsEqual(ExcelAddressBase a1, ExcelAddressBase a2)
-        {
-            return a1._fromRow == a2._fromRow &&
-                    a1._toRow == a2._toRow &&
-                    a1._fromCol == a2._fromCol &&
-                    a1._toCol == a2._toCol;
-        }
-
-    }
-    /// <summary>
     /// Range address with the address property readonly
     /// </summary>
     public class ExcelAddress : ExcelAddressBase
@@ -1640,180 +91,6 @@ namespace OfficeOpenXml
             {                
                 SetAddress(value, null, null);
                 ChangeAddress();
-            }
-        }
-    }
-    /// <summary>
-    /// Range address used in the formula parser
-    /// </summary>
-    public class ExcelFormulaAddress : ExcelAddressBase
-    {
-        /// <summary>
-        /// Creates a Address object
-        /// </summary>
-        internal ExcelFormulaAddress()
-            : base()
-        {
-        }
-
-        /// <summary>
-        /// Creates an Address object
-        /// </summary>
-        /// <param name="fromRow">start row</param>
-        /// <param name="fromCol">start column</param>
-        /// <param name="toRow">End row</param>
-        /// <param name="toColumn">End column</param>
-        public ExcelFormulaAddress(int fromRow, int fromCol, int toRow, int toColumn)
-            : base(fromRow, fromCol, toRow, toColumn)
-        {
-            _ws = "";
-        }
-        /// <summary>
-        /// Creates an Address object
-        /// </summary>
-        /// <param name="address">The formula address</param>
-        /// <param name="worksheet">The worksheet</param>
-        public ExcelFormulaAddress(string address, ExcelWorksheet worksheet)
-            : base(address, worksheet?.Workbook, worksheet?.Name)
-        {
-            SetFixed();
-        }
-        
-        internal ExcelFormulaAddress(string ws, string address)
-            : base(address)
-        {
-            if (string.IsNullOrEmpty(_ws)) _ws = ws;
-            SetFixed();
-        }
-        internal ExcelFormulaAddress(string ws, string address, bool isName)
-            : base(address, isName)
-        {
-            if (string.IsNullOrEmpty(_ws)) _ws = ws;
-            if(!isName)
-                SetFixed();
-        }
-
-        private void SetFixed()
-        {
-            if (Address.IndexOf('[') >= 0) return;
-            var address=FirstAddress;
-            if(_fromRow==_toRow && _fromCol==_toCol)
-            {
-                GetFixed(address, out _fromRowFixed, out _fromColFixed);
-            }
-            else
-            {
-                var cells = address.Split(':');                
-                if(cells.Length>1) //If 1 then the address is the entire worksheet
-                {
-                    GetFixed(cells[0], out _fromRowFixed, out _fromColFixed);
-                    GetFixed(cells[1], out _toRowFixed, out _toColFixed);
-                }
-            }
-        }
-
-        private void GetFixed(string address, out bool rowFixed, out bool colFixed)
-        {            
-            rowFixed=colFixed=false;
-            var ix=address.IndexOf('$');
-            while(ix>-1)
-            {
-                ix++;
-                if(ix < address.Length)
-                {
-                    if(address[ix]>='0' && address[ix]<='9')
-                    {
-                        rowFixed=true;
-                        break;
-                    }
-                    else
-                    {
-                        colFixed=true;
-                    }
-                }
-                ix = address.IndexOf('$', ix);
-            }
-        }
-        /// <summary>
-        /// The address for the range
-        /// </summary>
-        /// <remarks>Examples of addresses are "A1" "B1:C2" "A:A" "1:1" "A1:E2,G3:G5" </remarks>
-        public new string Address
-        {
-            get
-            {
-                if (string.IsNullOrEmpty(_address) && _fromRow>0)
-                {
-                    _address = GetAddress(_fromRow, _fromCol, _toRow, _toCol, _fromRowFixed, _toRowFixed, _fromColFixed, _toColFixed);
-                }
-                return _address;
-            }
-            set
-            {                
-                SetAddress(value, null, null);
-                ChangeAddress();
-                SetFixed();
-            }
-        }
-        internal new List<ExcelFormulaAddress> _addresses;
-        /// <summary>
-        /// Addresses can be separated by a comma. If the address contains multiple addresses this list contains them.
-        /// </summary>
-        public new List<ExcelFormulaAddress> Addresses
-        {
-            get
-            {
-                if (_addresses == null)
-                {
-                    _addresses = new List<ExcelFormulaAddress>();
-                }
-                return _addresses;
-
-            }
-        }
-        internal string GetOffset(int row, int column, bool withWbWs=false)
-        {
-            int fromRow = _fromRow, fromCol = _fromCol, toRow = _toRow, tocol = _toCol;
-            var isMulti = (fromRow != toRow || fromCol != tocol);
-            if (!_fromRowFixed)
-            {
-                fromRow += row;
-            }
-            if (!_fromColFixed)
-            {
-                fromCol += column;
-            }
-            if (isMulti)
-            {
-                if (!_toRowFixed)
-                {
-                    toRow += row;
-                }
-                if (!_toColFixed)
-                {
-                    tocol += column;
-                }
-            }
-            else
-            {
-                toRow = fromRow;
-                tocol = fromCol;
-            }
-            string a = GetAddress(fromRow, fromCol, toRow, tocol, _fromRowFixed, _fromColFixed, _toRowFixed, _toColFixed);
-            if (Addresses != null)
-            {
-                foreach (var sa in Addresses)
-                {
-                    a+="," + sa.GetOffset(row, column, withWbWs);
-                }
-            }
-            if(withWbWs)
-            {
-                return GetAddressWorkBookWorkSheet() + a;
-            }
-            else
-            {
-                return a;
             }
         }
     }
